requirements.txt
flask_sqlalchemy
OpenAlchemy
PyYAML
alembic
passlib
flake8
pytest
flask-mail
python-dotenv
itsdangerous